+++ src/core/films/section/tv.component.js
@@ -3,7 +3,7 @@ import { useHistory } from "react-router-dom";
 
 //own
 import useInitialiceTv from "../../main/use-initialice-tv";
-import { videosActions } from "../../../store/collections/movies/trailers/videos-slice";
+import { videosActions } from "../../../store/collections/movies/details/detail-slice";
 import Spinner from "../../../components/spinner";
 import Tv from "../tv";
