+++ src/core/movies/use-initialice-videos-tv.js
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
-
-import { videosActions } from "../../store/collections/movies/trailers/videos-slice";
+//own
+import { videosActions } from "../../store/collections/movies/details/detail-slice";
 import { getTvVideos } from "../../store/collections/tv/trailers/get-tv-videos";
 
 export default function useInitialiceTvVideos(id) {
