+++ src/store/collections/movies/details/detail-slice.js
