+++ src/store/index.js
@@ -1,9 +1,9 @@
 import { configureStore } from "@reduxjs/toolkit";
 import moviesSlice from "./collections/movies/movies-slice";
-import videosTvSlice from "./collections/tv/trailers/videos-tv-slice";
+import videosTvSlice from "./collections/tv/details/detail-tv-slice";
 import tvSlice from "./collections/tv/tv-slice";
 import genresSlice from "./genres/genres-slice";
-import videosSlice from "./collections/movies/trailers/videos-slice";
+import videosSlice from "./collections/movies/details/detail-slice";
 import searchSlice from "./search/search-slice";
 import topRatedSlice from "./collections/top/top-rated-slice";
